Remove function inlining. Fix api
